+++ Gemfile
@@ -55,7 +55,7 @@ gem 'slim-rails'
 gem 'devise', '~> 4.9.3'
 
 # gem 'loggable_activity', '~> 0.1.54', path: '/Users/maxgronlund/Documents/Max/open-source/LoggableActivity'
-gem 'loggable_activity', '~> 0.1.54'
+gem 'loggable_activity', '~> 0.1.55'
 
 # Use Active Storage variants [https://guides.rubyonrails.org/active_storage_overview.html#transforming-images]
 # gem "image_processing", "~> 1.2"
