+++ app/views/loggable_activity/templates/demo/journal/x_update.html.slim
@@ -4,7 +4,7 @@ tr
   td
     - update_activity_attrs = activity.update_activity_attrs
     - if update_attrs = update_activity_attrs[:update_attrs]
-      / b = update_attrs[:record_class]
+      / b = update_attrs[:record_type]
       / br/
       
       - update_attrs[:attrs].each do |attr|
