+++ app/views/loggable_activity/templates/shared/_update_attrs.html.slim
@@ -1,4 +1,4 @@
-b.text-secondary = update_relation_class(update_attrs)
+b.text-secondary = relation_type(update_attrs)
 br/
 - if update_attrs[:attrs] == :deleted
   = t('loggable.activity.deleted')
