+++ config/loggable_activity.yaml
@@ -1,5 +1,6 @@
 User: 
   record_display_name: full_name
+  route: demo_user_path
   loggable_attrs: 
     - first_name
     - last_name 
@@ -17,6 +18,7 @@ User:
         - sex
         - religion
     - belongs_to: :demo_address
+      route: demo_address_path
       model: Demo::Address
       loggable_attrs:
         - street
@@ -29,6 +31,7 @@ User:
         - name
 Demo::Address: 
   record_display_name: full_address
+  route: demo_address_path
   actor_display_name: email
   loggable_attrs:
     - street
@@ -51,6 +54,7 @@ Demo::Product:
     - destroy
 Demo::Club: 
   record_display_name: name 
+  route: demo_club_path
   loggable_attrs: 
     - name
   auto_log:
@@ -67,6 +71,7 @@ Demo::Club:
         - postal_code
 Demo::Journal: 
   record_display_name: title 
+  route: demo_journal_path
   loggable_attrs: 
     - title 
     - body
@@ -78,6 +83,7 @@ Demo::Journal:
   relations:
     - belongs_to: :doctor
       model: User 
+      route: demo_user_path
       data_owner: false 
       loggable_attrs:
         - first_name 
@@ -85,6 +91,7 @@ Demo::Journal:
         - email 
         - user_type
     - belongs_to: :patient
+      route: demo_user_path
       model: User 
       data_owner: true
       loggable_attrs:
