+++ spec/requests/users_spec.rb
@@ -59,7 +59,9 @@ RSpec.describe 'Users', type: :request do
       expect(activity.actor).to eq(user)
       expect(activity.attrs).to match([
                                         {
-                                          record_class: 'User',
+                                          record_type: 'User',
+                                          record_id: a_kind_of(Integer),
+                                          path: a_kind_of(String),
                                           payload_type: 'primary_payload',
                                           attrs: {
                                             'first_name' => a_kind_of(String),
